64-bit tsc
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@5411 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/arch.h
@@ -39,7 +39,7 @@ physaddr_t ppn2pa(ppn_t pn);
 extern char boot_cmdline[];
 void machine_reboot(void);
 uintptr_t karch_get_sp(void);
-uintptr_t karch_get_tsc(void);
+uint64_t karch_get_tsc(void);
 void karch_jmpbuf_init(struct jos_jmp_buf *jb, void *fn, void *stackbase);
 
 /*
